requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.3.0
pydantic-settings==2.0.3
email-validator==2.0.0
python-dotenv==1.0.0
azure-storage-blob==12.17.0
azure-identity==1.14.0
aiofiles==23.2.1
openai==0.28.0